# fix(perf): Filter queue in PackageHoister.seed instead of splicing