@arcanis 2018-08-09
Btw, I'm a bit curious: how did you manage to reach 2k dependencies at the top-level? Even here at FB we don't have that many. Knowing the use case could be useful!