@panrafal 2018-08-09
To be clear - we have 2k packages in `node_modules`, not in `package.json`.

Our frontend app is pretty big and kept in a single monorepo with ~200 packages. There's the testing platform, GraphQL server and tooling too - so it adds up pretty quickly.