@arcanis 2018-08-09
I see, thanks - I tried installing the top 1000 npm packages (which, counting transitive dependencies, makes for ~5k packages), and it installed in ~84s with a warm cache. There was a small pause before the linking started, but less than 10s.

Can you try running this command without your patch and see if you get different results? I'd like to understand under which circumstances these slowdowns happen, and how to reproduce it myself:

```
curl 'https://gist.githubusercontent.com/anvaka/8e8fa57c7ee1350e3491/raw/ab776d4b9015832fbc9035b3d302761408cf6de7/01.most-dependent-upon.md' | grep -F '. [' | grep -E -o '\[[^]]+\]' | grep -o -E '[^][]+' | xargs yarn add --ignore-scripts
```