@panrafal 2018-08-09
@arcanis I changed it to push to the new queue, as it's actually simpler (and does practically the same as calling `filter`).

As for recreating this, I timed all versions (1.7.0, 1.9.4, my fix) on our repo and... there's no lag on any one of them now. In the mean time I cleared the cache and reinstalled the modules. Yesterday I could recreate this no problem. Now it works alright 😕. So maybe there needs to be some perfect storm of state. We'll be running the modified version for now and I'll keep an eye if it ever happens again.