@arcanis 2018-08-09
Sounds good 🙂 I'll still merge it since the changes make sense. Even if we don't get perfs, it's still better engineering imo 👍